feat(api): unify character endpoints under /api/game/characters; deprecate /api/characters; normalize spawn + JSON errors
