+++ server/config.py
@@ -1,6 +1,9 @@
-START_POS = (12, 15)
+"""Shared gameplay coordinate constants."""
+
 # Demo gameplay coordinates
 START_TOWN_COORDS = (12, 15)
+# Legacy alias used by parts of the codebase; keep in sync with START_TOWN_COORDS
+START_POS = START_TOWN_COORDS
 PORT_TOWN_COORDS = (14, 9)
 AMBUSH_COORDS = (13, 12)
 TOWN_GRID_SIZE = (3, 3)
