+++ templates/item_forge.html
@@ -205,7 +205,7 @@
         <div style="font-size:11px; color:var(--muted)">Temper steel, bind runes, mint instances</div>
       </div>
     </div>
-    <span class="tag">/api/items • /api/item_instances • /api/characters/&lt;id&gt;/inventory</span>
+    <span class="tag">/api/items • /api/item_instances • /api/game/characters/&lt;id&gt;/inventory</span>
   </header>
 
   <main>
@@ -383,7 +383,7 @@
       mintInstance: (payload) =>
         fetch('/api/item_instances', { method:'POST', headers:{'Content-Type':'application/json'}, body:JSON.stringify(payload) }),
       grantInventory: (character_id, payload) =>
-        fetch('/api/characters/'+encodeURIComponent(character_id)+'/inventory', { method:'POST', headers:{'Content-Type':'application/json'}, body:JSON.stringify(payload) }),
+        fetch('/api/game/characters/'+encodeURIComponent(character_id)+'/inventory', { method:'POST', headers:{'Content-Type':'application/json'}, body:JSON.stringify(payload) }),
     };
 
     /* ---------- DOM ---------- */
@@ -563,7 +563,7 @@
 
         const c1 = `curl -s -X POST http://localhost:5000/api/items -H "Content-Type: application/json" -d '${JSON.stringify(payload)}'`;
         const c2 = `curl -s -X POST http://localhost:5000/api/item_instances -H "Content-Type: application/json" -d '${JSON.stringify({item_id:payload.item_id,item_version:payload.item_version,quantity:qty})}'`;
-        const c3 = `curl -s -X POST http://localhost:5000/api/characters/${cid}/inventory -H "Content-Type: application/json" -d '${JSON.stringify({slot_index:slot,item_id:payload.item_id,instance_id:"$INSTANCE_ID",qty:qty,equipped:eq})}'`;
+        const c3 = `curl -s -X POST http://localhost:5000/api/game/characters/${cid}/inventory -H "Content-Type: application/json" -d '${JSON.stringify({slot_index:slot,item_id:payload.item_id,instance_id:"$INSTANCE_ID",qty:qty,equipped:eq})}'`;
         return [c1, c2, c3].join('\n# INSTANCE_ID is value from previous response\n');
       }catch(e){ return '# invalid JSON/payload – cannot build cURL' }
     }
