+++ tests/test_character_location_persistence.py
@@ -1,5 +1,5 @@
 from app import create_app
-from server.config import START_POS
+from server.config import START_TOWN_COORDS
 import uuid
 
 
@@ -16,19 +16,19 @@ def test_cur_loc_updates_and_persists():
         )
         assert reg.status_code == 200
         # create and select character
-        created = client.post('/api/characters', json={'name': 'Hero', 'class_id': 'warrior'})
+        created = client.post('/api/game/characters', json={'name': 'Hero', 'class_id': 'warrior'})
         cid = created.get_json()['character_id']
-        client.post('/api/characters/select', json={'character_id': cid})
+        client.post('/api/game/characters/select', json={'character_id': cid})
         # spawn and move
         client.post('/api/spawn', json={})
         client.post('/api/move', json={'dx': 1, 'dy': 0})
         ch = db.session.get(Character, cid)
-        assert ch.x == START_POS[0] + 1
-        assert ch.y == START_POS[1]
+        assert ch.x == START_TOWN_COORDS[0] + 1
+        assert ch.y == START_TOWN_COORDS[1]
         assert ch.cur_loc == f"{ch.x},{ch.y}"
         # clear session and spawn again; position should persist
         with client.session_transaction() as sess:
             sess.clear()
         resp = client.post('/api/spawn', json={})
         data = resp.get_json()
-        assert data['player']['pos'] == [START_POS[0] + 1, START_POS[1]]
+        assert data['player']['pos'] == [START_TOWN_COORDS[0] + 1, START_TOWN_COORDS[1]]
